add fetch commands

+++ src/database.rs
@@ -28,7 +28,7 @@ pub async fn initialize_database() -> AsyncResult<()> {
   Ok(())
 }
 
-pub async fn get_all_birthdays(user_id: i32) -> AsyncResult<Vec<bth::Birthday>> {
+pub async fn get_all_birthdays(user_id: i64) -> AsyncResult<Vec<bth::Birthday>> {
     let mut birthdays: Vec<bth::Birthday> = Vec::new();
     let c = CONNECTION.lock().await;
     let mut statement = c.prepare(&format!("SELECT name, day, month, reminder, id FROM person WHERE user_id ={}", user_id))?;
@@ -48,19 +48,19 @@ pub async fn get_all_birthdays(user_id: i32) -> AsyncResult<Vec<bth::Birthday>> 
     Ok(birthdays)
 }
 
-pub async fn create_birthday(user_id: i32, birthday: &bth::Birthday) -> AsyncResult<()> {
+pub async fn create_birthday(user_id: i64, birthday: &bth::Birthday) -> AsyncResult<()> {
     CONNECTION.lock().await.execute("INSERT INTO person (user_id, name, day, month, reminder) VALUES (?1, ?2, ?3, ?4, ?5)", params![user_id, birthday.name, birthday.day, birthday.month, birthday.reminder],)?;
 
     Ok(())
 }
 
-pub async fn delete_birthday(user_id: i32, id: i32) -> AsyncResult<()>{
+pub async fn delete_birthday(user_id: i64, id: i64) -> AsyncResult<()>{
     CONNECTION.lock().await.execute("DELETE FROM person WHERE id=?1 AND user_id=?2", params![id,user_id],)?;
 
     Ok(())
 }
 
-pub async fn edit_birthday(user_id: i32, birthday: &bth::Birthday) -> AsyncResult<()>{
+pub async fn edit_birthday(user_id: i64, birthday: &bth::Birthday) -> AsyncResult<()>{
     CONNECTION.lock().await.execute("UPDATE person SET name = ?1, day = ?2, month = ?3, reminder=?4 WHERE id = ?5 AND user_id = ?6", params![birthday.name, birthday.day, birthday.month, birthday.reminder, birthday.id,user_id],)?;
 
     Ok(())
